requirements.txt
arrow==0.12.1
certifi==2018.8.24
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==2.1.1
django-filter==2.0.0
djangorestframework==3.8.2
httpie==0.9.9
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
Pillow==5.3.0
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.23

